Bug fix and finally making it work
Receiver now will listen for any packet.

ParsePacket will not crash the app if the user index is not found

Added the possibility to wait for updateOnlineUsers, there was an issue with joining the chat where the users would not be updated when the receiver was being instantiated resulting in a user not having its audio device.